app/requirements.txt
flask
requests
flask-cors
docker
flask
requests
flask-cors
docker
openai
pymongo
python-dotenv
google-generativeai
